Move thruster config structs into the algorithm header
